Make OpenTelemetry imports conditional
